Centralize metadata operand naming
Move the neutral IL operand naming contract into MetadataPrimitives and provide the single implementation from Metadata. Remove the repeated adapters from product, harness, and roundtrip consumers while preserving the peer dependency boundary.


Copilot-Session: b7300045-6e27-433f-9dfc-f51c4760452e
