+++ src/dotnet-inspect.Tests/MetadataInstructionProducer.cs
@@ -1,7 +1,6 @@
 using System.Reflection;
 using System.Reflection.Metadata;
 using System.Reflection.PortableExecutable;
-using DotnetInspector.Inspectors;
 using ILInspector.Instructions;
 using ILInspector.Metadata;
